wesago_app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.1
    # via
    #   black
    #   wesago
aiosignal==1.3.1
    # via
    #   aiohttp
    #   wesago
amqp==5.2.0
    # via
    #   kombu
    #   wesago
anyio==4.2.0
    # via
    #   wesago
    #   starlette
ariadne==0.21
    # via
    #   ariadne-django
    #   wesago
ariadne-django==0.3.0
    # via wesago
asgiref==3.7.2
    # via
    #   django
    #   wesago
attrs==23.1.0
    # via
    #   aiohttp
    #   wesago
billiard==4.2.0
    # via
    #   celery
    #   wesago
black==23.12.0
    # via wesago
celery[redis]==5.3.6
    # via wesago
certifi==2023.11.17
    # via
    #   wesago
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   wesago
charset-normalizer==3.3.2
    # via
    #   wesago
    #   requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   wesago
click-didyoumean==0.3.0
    # via
    #   celery
    #   wesago
click-plugins==1.1.1
    # via
    #   celery
    #   wesago
click-repl==0.3.0
    # via
    #   celery
    #   wesago
coverage[toml]==6.5.0
    # via
    #   coveralls
    #   wesago
    #   pytest-cov
coveralls==3.3.1
    # via wesago
cryptography==41.0.7
    # via
    #   wesago
    #   social-auth-core
defusedxml==0.8.0rc2
    # via
    #   wesago
    #   python3-openid
    #   social-auth-core
django==4.2.8
    # via
    #   ariadne-django
    #   django-anymail
    #   django-debug-toolbar
    #   django-js-asset
    #   django-simple-sso
    #   djangorestframework
    #   wesago
    #   social-auth-app-django
    #   webservices
django-anymail==10.2
    # via -r requirements.in
django-debug-toolbar==4.2.0
    # via wesago
django-js-asset==2.2.0
    # via
    #   django-mptt
    #   wesago
django-mptt==0.15.0
    # via wesago
django-simple-sso==1.2.0
    # via wesago
djangorestframework==3.14.0
    # via wesago
docopt==0.6.2
    # via
    #   coveralls
    #   wesago
faker==21.0.0
    # via wesago
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
    #   wesago
graphql-core==3.2.3
    # via
    #   ariadne
    #   wesago
html5lib==1.1
    # via wesago
idna==3.6
    # via
    #   anyio
    #   wesago
    #   requests
    #   yarl
iniconfig==2.0.0
    # via
    #   wesago
    #   pytest
itsdangerous==0.24
    # via
    #   django-simple-sso
    #   wesago
    #   webservices
kombu==5.3.4
    # via
    #   celery
    #   wesago
markdown==3.5.1
    # via wesago
# wesago==0.39.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   wesago
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   wesago
oauthlib==3.2.2
    # via
    #   wesago
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via
    #   black
    #   wesago
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   wesago
pillow==10.1.0
    # via wesago
platformdirs==4.1.0
    # via
    #   black
    #   wesago
pluggy==1.3.0
    # via
    #   wesago
    #   pytest
prompt-toolkit==3.0.43
    # via
    #   click-repl
    #   wesago
psycopg[binary]==3.1.16
    # via wesago
psycopg-binary==3.1.16
    # via
    #   wesago
    #   psycopg
pycparser==2.21
    # via
    #   cffi
    #   wesago
pyjwt==2.8.0
    # via
    #   wesago
    #   social-auth-core
pytest==7.4.3
    # via
    #   wesago
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
    #   syrupy
pytest-cov==4.1.0
    # via wesago
pytest-django==4.7.0
    # via wesago
pytest-mock==3.12.0
    # via wesago
python-dateutil==2.8.2
    # via
    #   celery
    #   faker
    #   wesago
python3-openid==3.2.0
    # via
    #   wesago
    #   social-auth-core
pytz==2023.3.post1
    # via
    #   djangorestframework
    #   wesago
pyyaml==6.0.1
    # via
    #   wesago
    #   responses
redis==5.0.1
    # via
    #   celery
    #   wesago
requests==2.31.0
    # via
    #   coveralls
    #   django-anymail
    #   wesago
    #   requests-oauthlib
    #   responses
    #   social-auth-core
    #   webservices
requests-oauthlib==1.3.1
    # via
    #   wesago
    #   social-auth-core
responses==0.24.1
    # via wesago
ruff==0.1.8
    # via wesago
sentry-sdk==1.39.1
    # via -r requirements.in
six==1.16.0
    # via
    #   html5lib
    #   wesago
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   wesago
social-auth-app-django==5.4.0
    # via wesago
social-auth-core==4.5.1
    # via
    #   wesago
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
    #   wesago
starlette==0.34.0
    # via
    #   ariadne
    #   wesago
syrupy==4.6.0
    # via wesago
typing-extensions==4.9.0
    # via
    #   ariadne
    #   wesago
    #   psycopg
tzdata==2023.3
    # via
    #   celery
    #   wesago
unidecode==1.3.7
    # via wesago
urllib3==2.1.0
    # via
    #   django-anymail
    #   wesago
    #   requests
    #   responses
    #   sentry-sdk
uwsgi==2.0.23
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
    #   wesago
wcwidth==0.2.12
    # via
    #   wesago
    #   prompt-toolkit
webencodings==0.5.1
    # via
    #   html5lib
    #   wesago
webservices[django]==0.7
    # via
    #   django-simple-sso
    #   wesago
yarl==1.9.4
    # via
    #   aiohttp
    #   wesago
